requirements.txt
numpy
scikit-allel
git+https://github.com/popsim-consortium/stdpopsim.git
